working on a second version of DrinkWater to test other approaches
